+++ examples/tracing_loops.py
@@ -37,7 +37,7 @@ plt.imshow(image, cmap="Spectral", origin="lower")
 # * The number of extra points in the loop below noise level to terminate a loop tracing ``ngap`` is 0.
 # * The base flux and median flux ratio ``qthresh1`` is 0.0
 # * The noise threshold in the image with repect to median flux ``qthresh2`` is 3.0 
-loops = trace.occult2(image, nsm1=3, rmin=30, lmin=25, nstruc=1000, nloop=1000, ngap=0, qthresh1=0.0, qthresh2=3.0, file=True)
+loops = trace.occult2(image, nsm1=3, rmin=30, lmin=25, nstruc=1000, nloop=1000, ngap=0, qthresh1=0.0, qthresh2=3.0, file=False)
 
 ###############################################################################
 # The `sunkit_image.trace.occult2` returns a list of all loop where each element
